linkedin_network/requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==43.0.1
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
gunicorn==23.0.0
PyJWT==2.9.0
packaging==24.1
psycopg2-binary==2.9.9
pycparser==2.22
sqlparse==0.5.1
django-cors-headers==4.4.0
pandas==2.2.0